Split find_inputs.py and more into separate module. Add utils module
