+++ batch_seqr_loader/utils.py
@@ -1,0 +1,29 @@
+"""
+Utility functions for the seqr loader
+"""
+
+import os
+from google.cloud import storage
+
+
+def file_exists(path: str) -> bool:
+    """
+    Check if the object exists, where the object can be:
+        * local file
+        * local directory
+        * Google Storage object
+        * Google Storage URL representing a *.mt or *.ht Hail data,
+          in which case it will check for the existence of a
+          *.mt/_SUCCESS or *.ht/_SUCCESS file.
+    :param path: path to the file/directory/object/mt/ht
+    :return: True if the object exists
+    """
+    if path.startswith('gs://'):
+        bucket = path.replace('gs://', '').split('/')[0]
+        path = path.replace('gs://', '').split('/', maxsplit=1)[1]
+        path = path.rstrip('/')  # '.mt/' -> '.mt'
+        if any(path.endswith(f'.{suf}') for suf in ['mt', 'ht']):
+            path = os.path.join(path, '_SUCCESS')
+        gs = storage.Client()
+        return gs.get_bucket(bucket).get_blob(path)
+    return os.path.exists(path)
